backend/requirements.txt
fastapi
uvicorn
python-multipart
biopython
numpy
scipy